+++ CHANGELOG.md
@@ -11,13 +11,19 @@ as well as to [Module version numbering](https://go.dev/doc/modules/version-numb
 ### Added
 
 - Add `color.ReportFlow` which is an extension of `middleware.ReportFlow`.
+- Add `otelgoyek.Middleware` and `otelgoyek.ExecutorMiddleware` which add
+  OpenTelemetry tracing instrumentation.
 
 ### Changed
 
 - Bump `github.com/goyek/goyek` to `2.2.0`.
 - Bump `github.com/fatih/color` to `1.17.0`.
 
-## [0.1.7](https://github.com/goyek/x/releases/tag/v0.1.7) - 2024-01-17
+### Remove
+
+- Drop support for Go 1.17, 1.18, 1.19, 1.20.
+
+## [0.1.7](https://github.com/goyek/x\/releases/tag/v0.1.7) - 2024-01-17
 
 ### Added